feat(bcs-bscp): feedserver support sync download file api
