+++ .gitignore
@@ -9,3 +9,4 @@ cover.out
 .envrc
 .vscode
 /bin
+/debug
