+++ cmd/bscp/common.go
@@ -61,13 +61,17 @@ var (
 
 	// commonEnvs variable definition
 	commonEnvs = map[string]string{
-		"biz":         "biz",
-		"app":         "app",
-		"labels":      "labels",
-		"labels_file": "labels_file",
-		"feed_addrs":  "feed_addrs",
-		"token":       "token",
-		"temp_dir":    "temp_dir",
+		"biz":            "biz",
+		"app":            "app",
+		"labels":         "labels",
+		"labels_file":    "labels_file",
+		"feed_addrs":     "feed_addrs",
+		"token":          "token",
+		"temp_dir":       "temp_dir",
+		"agent_id":       "agent_id",
+		"cluster_id":     "cluster_id",
+		"pod_id":         "pod_id",
+		"container_name": "container_name",
 	}
 
 	watchEnvs = map[string]string{
